requirements.txt
anthropic==0.14.0
streamlit==1.45.1
notion-client==2.2.1
requests==2.31.0
python-dotenv==1.0.0 